@@ -1,24 +1,21 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
-apply plugin: 'io.fabric'
+plugins {
+    id 'com.android.application'
+    id 'kotlin-android'
+    id 'com.google.gms.google-services'
+    id 'com.google.firebase.crashlytics'
+}
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion '29.0.3'
+    compileSdkVersion 30
+    buildToolsVersion '31.0.0'
 
     defaultConfig {
         applicationId 'com.henrykvdb.sttt'
         minSdkVersion 21
-        targetSdkVersion 29
+        targetSdkVersion 30
         versionCode 28
         versionName '2.2.2'
         resConfigs "en"
-    }
-
-    compileOptions {
-        sourceCompatibility 1.8
-        targetCompatibility 1.8
     }
 
     buildTypes {
@@ -30,25 +27,30 @@ android {
         }
     }
 
-    sourceSets {
-        main {
-            kotlin.srcDirs = ["src/main"]
-            java.srcDirs = ["src/main"]
-            resources.srcDirs = ["src/main"]
-        }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
+
+    buildFeatures {
+        viewBinding true
     }
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.github.AppIntro:AppIntro:4.2.3'
+    implementation 'androidx.core:core-ktx:1.6.0'
+    implementation 'androidx.appcompat:appcompat:1.3.1'
+    implementation 'com.google.android.material:material:1.4.0'
 
-    implementation 'com.google.android.material:material:1.1.0'
-    implementation 'com.google.firebase:firebase-core:17.4.4'
-    implementation 'com.google.firebase:firebase-ads:19.2.0'
-    implementation 'com.crashlytics.sdk.android:crashlytics:2.10.1'
+    implementation files('libs/SuperTTTApi.jar')
+    implementation 'com.github.AppIntro:AppIntro:6.1.0'
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation platform('com.google.firebase:firebase-bom:28.3.0')
+    implementation 'com.google.firebase:firebase-analytics-ktx'
+    implementation 'com.google.firebase:firebase-crashlytics-ktx'
+    implementation 'com.google.firebase:firebase-ads:20.2.0'
 }
-
-apply plugin: 'com.google.gms.google-services'
